Update README.md
This is a minecraft modpack, for minecraft version 1.7.10, and will be mainly focused around GregTech which is a mod that greatly enhances and changes minecrafts normal progression, and adds many tech related items/machines/multiblocks etc. This modpack will include a series of Chapters and quests to guide you through the modpack.